+++ R/vctrs-math.R
@@ -42,9 +42,9 @@ vec_math_bigfloat <- function(.fn, .x, ..., na.rm = FALSE) {
 
     # Other
     mean = c_bigfloat_sum(.x, na.rm) / sum(!is.na(.x)),
-    is.nan = is.nan(allow_lossy_cast(vec_cast(.x, double()))),
-    is.infinite = is.infinite(allow_lossy_cast(vec_cast(.x, double()))),
-    is.finite = is.finite(allow_lossy_cast(vec_cast(.x, double()))),
+    is.nan = vec_data(.x) %|% "NA" == "NaN",
+    is.infinite = vec_data(.x) %in% c("Inf", "-Inf"),
+    is.finite = !(vec_data(.x) %in% c(NA, "NaN", "Inf", "-Inf")),
 
     # else
     stop_unsupported(.x, .fn)
@@ -77,9 +77,9 @@ vec_math.bignum_biginteger <- function(.fn, .x, ..., na.rm = FALSE) {
 
     # Other
     mean = c_biginteger_sum(.x, na.rm) / sum(!is.na(.x)),
-    is.nan = rep_len(FALSE, length(.x)),
+    is.nan = rep_along(.x, FALSE),
     is.finite = !is.na(.x),
-    is.infinite = rep_len(FALSE, length(.x)),
+    is.infinite = rep_along(.x, FALSE),
 
     # else
     vec_math_bigfloat(.fn, .x, ..., na.rm = na.rm)